requirements.txt
pillow
numpy
streamlit
matplotlib
pandas